Level up!!! and exp
Chimeras now gain exp from dungeon clear - currently only based on the difficulty of the dungeon, we should probably adjust to be based on # killed. overcompensation of currency bug fixed. Leveling up affects health and attack of chimeras, and health only of monsters. Monster levels are based solely on the dungeon level they're on rn